refactor(client/tunnel): convert tunnel_* files into tunnel package module
