--- argus/client/tunnel.py
@@ -1,13 +1,0 @@
-from argus.client.tunnel_api import resolve_tunnel_config, resolve_tunnel_config_with_reason
-from argus.client.tunnel_models import TunnelClientError, TunnelConfig
-from argus.client.tunnel_ssh import SSHTunnel
-from argus.client.tunnel_state import delete_cached_tunnel_state
-
-__all__ = [
-    "SSHTunnel",
-    "TunnelClientError",
-    "TunnelConfig",
-    "delete_cached_tunnel_state",
-    "resolve_tunnel_config",
-    "resolve_tunnel_config_with_reason",
-]
