+++ argus/client/tunnel/__init__.py
@@ -1,0 +1,13 @@
+from argus.client.tunnel.api import resolve_tunnel_config, resolve_tunnel_config_with_reason
+from argus.client.tunnel.models import TunnelClientError, TunnelConfig
+from argus.client.tunnel.ssh import SSHTunnel
+from argus.client.tunnel.state import delete_cached_tunnel_state
+
+__all__ = [
+    "SSHTunnel",
+    "TunnelClientError",
+    "TunnelConfig",
+    "delete_cached_tunnel_state",
+    "resolve_tunnel_config",
+    "resolve_tunnel_config_with_reason",
+]
